requirements.txt
# Framework Web FastAPI
fastapi==0.104.1  # Inclui starlette para middleware de sessões

# Servidor ASGI
uvicorn[standard]==0.24.0

# Validação de dados e modelos
pydantic==2.5.0
pydantic[email]
pydantic-settings==2.1.0

# Requisições HTTP
requests==2.31.0
requests-oauthlib==1.3.1

# Automação Web (Selenium)
selenium==4.15.2
webdriver-manager==4.0.2

# Templates HTML
jinja2==3.1.2

# Gerenciamento de variáveis de ambiente
python-dotenv==1.0.0

# Google Drive API
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-httplib2==0.1.1

# Webapp dependencies
python-multipart==0.0.6  # Upload de arquivos (planilhas)
aiofiles==23.2.1  # Operações assíncronas de arquivo
itsdangerous==2.1.2  # Gerenciamento de sessões
openpyxl==3.1.2  # Processamento de planilhas Excel
email-validator==2.1.0  # Validação de emails
